Rename `ContextExtension` into `Plugin` and add doc
